clockintr: rename "clockintr_schedule_locked" to "clockqueue_pend_insert"
The function clockintr_schedule_locked() is poorly named.  "Schedule" is
a broad word that connotes too much and "clockintr" implies that we are
mainly operating on the clockintr itself, which is untrue.

clockqueue_pend_insert() is a better name: it is clearer that we are
inserting an element into cq_pend.

(cherry picked from commit 9bc72213b6cbaf5b29182fe4cf3edf08ef4a64a7)