vmd(8): fix vm pause deadlock.
When vcpu threads pause, they are holding the run mutex lock. If
the event thread is asked to assert an irq on the pic and interrupts
are pending, it will try to take the run mutex lock on the vcpu.
This deadlocks.

Release the lock in the vcpu thread before waiting on the pause
condition variable.

ok mlarkin@

(cherry picked from commit 71ab85de86c9965af0a2d6132e10736e2dd4e37f)